requirements.txt
aniso8601==9.0.1
cachetools==5.2.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
google-auth==2.13.0
google-auth-oauthlib==0.5.3
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.42
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1